Patch-ioner/src/denseclip/clip_loader/requirements.txt
# Requirements for DenseCLIP to CLIP Loader
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
PyYAML>=5.4.0
numpy>=1.20.0
regex>=2021.4.4
